Move signals to scos_actions and ran pre-commit.
