+++ scos_actions/actions/acquire_stepped_freq_tdomain_iq.py
@@ -46,9 +46,9 @@ from scos_actions.actions.acquire_single_freq_tdomain_iq import (
     NUM_SKIP,
     SingleFrequencyTimeDomainIqAcquisition,
 )
-from scos_actions.actions.interfaces.signals import measurement_action_completed
 from scos_actions.hardware.mocks.mock_gps import MockGPS
 from scos_actions.metadata.sigmf_builder import Domain, MeasurementType
+from scos_actions.signals import measurement_action_completed
 from scos_actions.utils import get_parameter
 
 logger = logging.getLogger(__name__)
