+++ scos_actions/actions/tests/test_stepped_freq_tdomain_iq.py
@@ -1,6 +1,5 @@
-from scos_actions.actions.interfaces.signals import measurement_action_completed
-from scos_actions.actions.tests.utils import SENSOR_DEFINITION, check_metadata_fields
 from scos_actions.discover import test_actions as actions
+from scos_actions.signals import measurement_action_completed
 
 SINGLE_TIMEDOMAIN_IQ_MULTI_RECORDING_ACQUISITION = {
     "name": "test_multirec_acq",
