+++ scos_actions/configs/sensor_calibration_example.json
