+++ scos_actions/configs/sigan_calibration_example.json
