+++ scos_actions/discover/__init__.py
@@ -1,5 +1,4 @@
 from scos_actions.actions import action_classes
-from scos_actions.actions.interfaces.signals import register_component_with_status
 from scos_actions.actions.logger import Logger
 from scos_actions.actions.monitor_sigan import MonitorSignalAnalyzer
 from scos_actions.actions.sync_gps import SyncGps
@@ -7,6 +6,7 @@ from scos_actions.discover.yaml import load_from_yaml
 from scos_actions.hardware.mocks.mock_gps import MockGPS
 from scos_actions.hardware.mocks.mock_sigan import MockSignalAnalyzer
 from scos_actions.settings import ACTION_DEFINITIONS_DIR, MOCK_SIGAN
+from scos_actions.signals import register_component_with_status
 
 mock_sigan = MockSignalAnalyzer(randomize_values=True)
 mock_gps = MockGPS()
