+++ scos_actions/hardware/__init__.py
@@ -6,7 +6,6 @@ from its_preselector.configuration_exception import ConfigurationException
 from its_preselector.controlbyweb_web_relay import ControlByWebWebRelay
 
 from scos_actions import utils
-from scos_actions.actions.interfaces.signals import register_component_with_status
 from scos_actions.capabilities import capabilities
 from scos_actions.hardware.mocks.mock_gps import MockGPS
 from scos_actions.settings import (
@@ -15,6 +14,7 @@ from scos_actions.settings import (
     PRESELECTOR_MODULE,
     SWITCH_CONFIGS_DIR,
 )
+from scos_actions.signals import register_component_with_status
 from scos_actions.status.status_registration_handler import status_registration_handler
 
 logger = logging.getLogger(__name__)
